service_net::interface

Struct VirtioNetDevice

pub struct VirtioNetDevice {
    rx_queue: VecDeque<Box<[u8]>>,
    guest_rx_queue: VecDeque<Box<[u8]>>,
    guest_mac: Option<[u8; 6]>,
}

smoltcp Device implementation backed by a kernel IPC frame queue.

Fields
rx_queue: VecDeque<Box<[u8]>>guest_rx_queue: VecDeque<Box<[u8]>>

Frames destined for the hypervisor guest, separated by dst MAC.

guest_mac: Option<[u8; 6]>
Implementations

impl VirtioNetDevice

pub fn new() -> Self

pub fn push_rx(&mut self, frame: Box<[u8]>)

Enqueue an inbound frame received from the kernel VirtIO net driver.

pub fn set_guest_mac(&mut self, mac: [u8; 6])

Register the guest MAC address for L2 bridging.

pub fn pop_guest_rx(&mut self) -> Option<Box<[u8]>>

Pop one frame from the guest RX queue.

pub fn pump_rx(&mut self) -> usize

Drain pending RX frames from the active NIC into the local queue.

Routes to the e1000 Driver Cell when registered; falls back to the kernel VirtIO NetRx syscall. Returns the number of frames pulled.

pub fn pump_rx_split(&mut self)

Drain pending RX frames, splitting by dst MAC when a guest MAC is registered.

pub fn get_driver_mac(&self) -> Option<[u8; 6]>

Query the e1000 Driver Cell for the MAC address, if registered.